move calculate_rsa_ckaid from secrets.c to rsapub.c
